More reliable testsuite - replace Jasmine with Chai
Jasmine was quietly not running all the tests, and when the server was
offline the integration tests did not fail, it just exited quietly when
run as a suite. When run individually it might fail, but lots of stack
traces for each expectation. Mocha / Chai reports the failure of the
'before' once, and it is easy to see that the other test is still
passing (and run).

Normally I don't like seeing green tests, but after not being sure
whether tests were run, I'm happy with it (can probably be switched with
a different reporter).

Now we get this output when the server is down:
stub server
    1) "before each" hook

  Request Builder
    ✓ can build a POST stub

  1 passing (36ms)
  1 failing

  1) stub server "before each" hook:
     Uncaught AssertionError: expected [Error: connect ECONNREFUSED 127.0.0.1:3002] to equal null
      at Request._callback (spec/node-spec.js:18:15)
      at self.callback (node_modules/request/request.js:198:22)
      at Request.onRequestError (node_modules/request/request.js:820:8)
      at Socket.socketErrorListener (_http_client.js:259:9)
      at emitErrorNT (net.js:1253:8)